modify: 피드 컨텐츠/댓글 컴포넌트화
